Move JSLibraries into Stack Packs
