+++ lighthouse-core/test/results/sample_v2.json
@@ -5626,5 +5626,6 @@
         "categoryGroups[a11y-tables-lists].description"
       ]
     }
-  }
+  },
+  "stackPacks": []
 }
